+++ R/gg_partial.R
@@ -113,5 +113,6 @@ gg_partial <- function(part_dta,
 
   result <- list(continuous = continuous, categorical = categorical)
   class(result) <- "gg_partial"
+  result <- .set_provenance(result, part_dta)
   return(result)
 }